Trait pallet_dynamic_fee::pallet::Config

pub trait Config: Config {
    type MinGasPriceBoundDivisor: Get<U256>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MinGasPriceBoundDivisor: Get<U256>

Bound divisor for min gas price.

Object Safety

This trait is not object safe.